Moved @@update-completion-percentage to deployment package.
