+++ src/euphorie/client/browser/webhelpers.py
@@ -936,44 +936,3 @@ class HelpMenu(WebHelpers):
 
     def __call__(self):
         return self.index()
-
-
-class UpdateCompletionPercentage(WebHelpers):
-    """ Utility view to fill in missing values for completion_percentage """
-
-    def log(self, entry):
-        logger.info(entry)
-        self._log = "\n".join((getattr(self, "_log", ""), entry))
-
-    def get_log(self):
-        return getattr(self, "_log", "")
-
-    def next_b_start(self):
-        if "overwrite" in self.request.form:
-            return self.request.get("b_start", 0) + self.request.get("b_size", 1000)
-        else:
-            return 0
-
-    def __call__(self):
-        if self.request.method == "POST":
-            self.log("Updating completion_percentage")
-            b_size = self.request.get("b_size", 1000)
-            b_start = self.request.get("b_start", 0)
-            self.log("Limiting to {} sessions; starting at {}".format(b_size, b_start))
-            query = Session.query(SurveySession)
-            if "overwrite" not in self.request.form:
-                query = query.filter(SurveySession.completion_percentage == None)
-                self.log("Ignoring sessions with existing non-null values")
-            else:
-                self.log("Overwriting existing non-null values")
-            query = (
-                query
-                .order_by(SurveySession.modified.desc())
-                .offset(b_start)
-                .limit(b_size)
-            )
-            self.log("Found {} sessions".format(query.count()))
-            for session in query:
-                self.update_completion_percentage(session)
-            self.log("Done")
-        return self.index()
