+++ src/euphorie/deployment/browser/templates/update-completion-percentage.pt
@@ -6,7 +6,6 @@
       xmlns:tal="http://xml.zope.org/namespaces/tal"
       meta:interpolation="true"
       i18n:domain="euphorie"
-      tal:define="client nocall:request/client; webhelpers nocall:context/@@webhelpers"
       >
 
 <body>
